@@ -17,6 +17,6 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=7.0"
+        "php": ">=8.0"
     }
 }
